+++ CHANGELOG.md
@@ -2,17 +2,26 @@
 
 All notable changes to Editorial Calendar.io will be documented in this file.
 
-## 1.0 - 3/15/2022
+## 1.0.3-1.0.4 - 3/31/2022
 
-LAUNCH!
+### Fixed
+
+- Trash posts date error
+- PostLinks hidden during update on all posts
 
 ## 1.0.2 - 3/30/2022
 
 ### Fixed
+
 - New Post button provides loading feedback
 - Fixed issues with shifting and unshifting time zones
 
 ## 1.0.1 - 3/15/2022 - IN PROGRESS
 
 ### Changed
+
 - package.json dependency updates (codacy/snyk)
+
+## 1.0 - 3/15/2022
+
+- LAUNCH!